Create pylearn2_cpu_model_function_builder.py
Adding script to load a GPU-trained pylearn2 model pickle on a computer with a CUDA compatible GPU and output a pickle of a theano function corresponding to a forward pass through the model (i.e. prediction function) on a *batch* of input points which importantly is compiled to CPU compatible code and so can be run on a non-GPU equipped computer.